[MISC] make hash public API and move them into own header files
