+++ include/seqan3/io/alignment_file/header.hpp
@@ -17,9 +17,9 @@
 #include <vector>
 
 #include <seqan3/alphabet/concept.hpp>
-#include <seqan3/alphabet/detail/hash.hpp>
 #include <seqan3/core/type_traits/pre.hpp>
 #include <seqan3/io/alignment_file/detail.hpp>
+#include <seqan3/range/hash.hpp>
 #include <seqan3/range/views/view_all.hpp>
 #include <seqan3/std/ranges>
