+++ test/unit/alphabet/alphabet_hash_test.cpp
@@ -11,6 +11,7 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/alphabet/quality/phred42.hpp>
 #include <seqan3/alphabet/quality/qualified.hpp>
+#include <seqan3/range/hash.hpp>
 
 using namespace seqan3;
